# Backend/docs/openapi-final.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIGGAB API",
    "version": "2.1.0",
    "description": "Sistema Gestor de Ganado Bovino (SIGGAB) - API REST.\n\nTodos los endpoints (excepto login y refresh) requieren Bearer token.\nRespuesta estandar: { success, data, message, errors }.",
    "contact": {
      "name": "SIGGAB Dev Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido en POST /auth/login"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "nullable": true
          },
          "message": {
            "type": "string",
            "example": "Operacion exitosa"
          },
          "errors": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "errors"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/produccion-leche": {
      "get": {
        "tags": [
          "ProductivoLeche"
        ],
        "summary": "Listar registros de produccion de leche",
        "parameters": [
          {
            "in": "query",
            "name": "idAnimal",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "idLote",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "APROBADO",
                "RECHAZADO"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registros obtenidos"
          }
        }
      },
      "post": {
        "tags": [
          "ProductivoLeche"
        ],
        "summary": "Registrar produccion de leche",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idAnimal",
                  "idLote",
                  "litrosProducidos",
                  "fechaRegistro"
                ],
                "properties": {
                  "idAnimal": {
                    "type": "integer"
                  },
                  "idLote": {
                    "type": "integer"
                  },
                  "litrosProducidos": {
                    "type": "number"
                  },
                  "fechaRegistro": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registro creado"
          },
          "400": {
            "description": "Error de validacion"
          }
        }
      }
    },
    "/produccion-leche/{id}": {
      "get": {
        "tags": [
          "ProductivoLeche"
        ],
        "summary": "Obtener registro de produccion por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro encontrado"
          },
          "404": {
            "description": "Registro no encontrado"
          }
        }
      },
      "patch": {
        "tags": [
          "ProductivoLeche"
        ],
        "summary": "Modificar registro de produccion de leche",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "litrosProducidos": {
                    "type": "number"
                  },
                  "fechaRegistro": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registro actualizado"
          },
          "400": {
            "description": "Error de validacion"
          },
          "404": {
            "description": "Registro no encontrado"
          }
        }
      }
    },
    "/produccion-leche/{id}/validar": {
      "patch": {
        "tags": [
          "ProductivoLeche"
        ],
        "summary": "Validar registro de produccion de leche",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "estadoValidacion"
                ],
                "properties": {
                  "estadoValidacion": {
                    "type": "string",
                    "enum": [
                      "APROBADO",
                      "RECHAZADO"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registro validado"
          },
          "400": {
            "description": "Error de validacion"
          },
          "404": {
            "description": "Registro no encontrado"
          }
        }
      }
    },
...
    "/reportes/sanitario": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Generar reporte sanitario",
        "parameters": [
          {
            "in": "query",
            "name": "fechaInicio",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "fechaFin",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "estadoAprobacion",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "APROBADO",
                "RECHAZADO"
              ]
            }
          },
          {
            "in": "query",
            "name": "idTipoEvento",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte generado (JSON, CSV o PDF)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/reportes/productivo": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Generar reporte productivo",
        "parameters": [
          {
            "in": "query",
            "name": "fechaInicio",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "fechaFin",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "idAnimal",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "idLote",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte generado (JSON, CSV o PDF)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/reportes/administrativo": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Generar reporte administrativo",
        "parameters": [
          {
            "in": "query",
            "name": "fechaInicio",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "fechaFin",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte generado (JSON, CSV o PDF)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/reportes/comparativo": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Generar reporte comparativo historico (RF13)",
        "parameters": [
          {
            "in": "query",
            "name": "modulo",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "sanitario",
                "productivo"
              ]
            }
          },
          {
            "in": "query",
            "name": "periodoAInicio",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "periodoAFin",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "periodoBInicio",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "periodoBFin",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte comparativo generado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}